API-11015: splitted declarations for Customer and Agent API
